requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-dotenv==1.4.2
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.1
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
style==1.1.0
update==0.0.1
uritemplate==4.1.1
